agent-openai/requirements.txt
openai>=1.35.0
python-dotenv>=1.0.0
pyyaml>=6.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
tqdm>=4.66.0
